layout/style/test/test_bug716226.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=716226
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 716226</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style id="s">
    @keyframes foo { }
  </style>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=716226">Mozilla Bug 716226</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 716226 **/
var sheet = $("s").sheet;
var rules = sheet.cssRules;
is(rules.length, 1, "Should have one keyframes rule");
var keyframesRule = rules[0];
var keyframeRules = keyframesRule.cssRules;
is(keyframeRules.length, 0, "Should have no keyframe rules yet");

keyframesRule.insertRule('0% { }');
is(keyframeRules.length, 1, "Should have a keyframe rule now");
var keyframeRule = keyframeRules[0];
is(keyframeRule.parentRule, keyframesRule,
   "Parent of keyframe should be keyframes");
is(keyframeRule.parentStyleSheet, sheet,
   "Parent stylesheet of keyframe should be our sheet");

is(keyframeRule.style.cssText, "", "Should have no declarations yet");
// Note: purposefully non-canonical cssText string so we can make sure we
// really invoked the CSS parser and serializer.
keyframeRule.style.cssText = "color:green";
is(keyframeRule.style.cssText, "color: green;",
   "Should have the declarations we set now");



</script>
</pre>
</body>
</html>
